requirements.txt
flask
flask-bcrypt
flask-login
flask-mail
flask-sqlalchemy
flask-wtf
itsdangerous
numpy
Pillow
wtforms
